pyproject.toml
[project]
name = "github-batch-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow[amazon,postgres]>=3.1.7",
]

[dependency-groups]
dev = [
    "ty>=0.0.20",
]
